Moved NWTree to core/tree.py
